Refactor PoolJoin into AddLiquidity
