+++ src/entities/join/index.ts
@@ -1,2 +1,2 @@
-export * from './poolJoin';
+export * from './addLiquidity';
 export * from './types';
